sonar cloud bug fixes for app module casting fixes,

Related work items: #16086

+++ app/src/main/java/a75f/io/renatus/util/Scroller.java
@@ -515,7 +515,7 @@ public class Scroller {
      */
     public void setFinalX(int newX) {
         mFinalX = newX;
-        mDeltaX = mFinalX - mStartX;
+        mDeltaX = (float) mFinalX - mStartX;
         mFinished = false;
     }
 
@@ -528,7 +528,7 @@ public class Scroller {
      */
     public void setFinalY(int newY) {
         mFinalY = newY;
-        mDeltaY = mFinalY - mStartY;
+        mDeltaY = (float) mFinalY - mStartY;
         mFinished = false;
     }
 
@@ -536,8 +536,8 @@ public class Scroller {
      * @hide
      */
     public boolean isScrollingInDirection(float xvel, float yvel) {
-        return !mFinished && Math.signum(xvel) == Math.signum(mFinalX - mStartX) &&
-                Math.signum(yvel) == Math.signum(mFinalY - mStartY);
+        return !mFinished && Math.signum(xvel) == Math.signum((float) mFinalX - mStartX) &&
+                Math.signum(yvel) == Math.signum((float) mFinalY - mStartY);
     }
 
     static class ViscousFluidInterpolator implements Interpolator {
